// apps/hermitboard-mobile-interface/src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: any;
  SessionToken: any;
  Time: any;
  Void: any;
};
// ...
export type AccountConnection = {
  __typename?: 'AccountConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<AccountEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type AccountEdge = {
  __typename?: 'AccountEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor'];
  /** The item at the end of the edge. */
  node?: Maybe<Account>;
};
// ...
export type AssetClass = Node & {
  __typename?: 'AssetClass';
  createdAt: Scalars['Time'];
  deletedAt?: Maybe<Scalars['Time']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  updatedAt: Scalars['Time'];
  value: AssetClassValue;
};
// ...
export enum AssetClassValue {
  CASH_OR_CASH_EQUIVALENT = 'CASH_OR_CASH_EQUIVALENT',
  COMMODITY = 'COMMODITY',
  CRYPTOCURRENCY = 'CRYPTOCURRENCY',
  EQUITY = 'EQUITY',
  FIXED_INCOME = 'FIXED_INCOME',
  FUTURE = 'FUTURE',
  REAL_ESTATE = 'REAL_ESTATE'
}
// ...
export type AuthRole = Node & {
  __typename?: 'AuthRole';
  accounts: AccountConnection;
  createdAt: Scalars['Time'];
  deletedAt?: Maybe<Scalars['Time']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  staffAccounts: StaffAccountConnection;
  updatedAt: Scalars['Time'];
  value: AuthRoleValue;
};
// ...
export enum AuthRoleValue {
  ADMIN = 'ADMIN',
  DEMO = 'DEMO',
  ENTERPRISE = 'ENTERPRISE',
  FREE = 'FREE',
  PLUS = 'PLUS',
  PRO = 'PRO',
  SUPER_ADMIN = 'SUPER_ADMIN',
  SUPPORT = 'SUPPORT'
}
// ...
export type AuthType = Node & {
  __typename?: 'AuthType';
  createdAt: Scalars['Time'];
  deletedAt?: Maybe<Scalars['Time']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  updatedAt: Scalars['Time'];
  value: AuthTypeValue;
};
// ...
export enum AuthTypeValue {
  APPLE = 'APPLE',
  FACEBOOK = 'FACEBOOK',
  GOOGLE = 'GOOGLE',
  LOCAL = 'LOCAL'
}
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['Cursor']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['Cursor']>;
};
// ...
export type SourceConnection = {
  __typename?: 'SourceConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<SourceEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type SourceEdge = {
  __typename?: 'SourceEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor'];
  /** The item at the end of the edge. */
  node?: Maybe<Source>;
};
// ...
export type SourceType = Node & {
  __typename?: 'SourceType';
  createdAt: Scalars['Time'];
  deletedAt?: Maybe<Scalars['Time']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  sources: SourceConnection;
  updatedAt: Scalars['Time'];
  value: SourceTypeValue;
};
// ...
export enum SourceTypeValue {
  BANK = 'BANK',
  CRYPTOCURRENCY_WALLET = 'CRYPTOCURRENCY_WALLET',
  DECENTRALIZED_EXCHANGE = 'DECENTRALIZED_EXCHANGE',
  EXCHANGE = 'EXCHANGE'
}
// ...
export type StaffAccountConnection = {
  __typename?: 'StaffAccountConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<StaffAccountEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type StaffAccountEdge = {
  __typename?: 'StaffAccountEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor'];
  /** The item at the end of the edge. */
  node?: Maybe<StaffAccount>;
};
// ...
export type TransactionType = Node & {
  __typename?: 'TransactionType';
  createdAt: Scalars['Time'];
  deletedAt?: Maybe<Scalars['Time']>;
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  updatedAt: Scalars['Time'];
  value: TransactionTypeValue;
};
// ...
export enum TransactionTypeValue {
  BUY = 'BUY',
  DIVIDEND_INCOME = 'DIVIDEND_INCOME',
  RENT_INCOME = 'RENT_INCOME',
  RENT_PAYMENT = 'RENT_PAYMENT',
  SELL = 'SELL',
  STAKE = 'STAKE',
  STOCK_DIVIDEND = 'STOCK_DIVIDEND',
  SWAP = 'SWAP'
}